[Glimmer2] Migrate class binding tests
